requirements.txt
flask==3.0.0
openai==1.3.0
requests==2.31.0
python-dotenv==1.0.0
apscheduler==3.10.1
python-dateutil==2.8.2
beautifulsoup4==4.12.2
html5lib==1.1
gunicorn==21.2.0
PyPDF2==3.0.1
PyGithub==2.1.1